Refactor: Use Double instead of Float for number types
This commit changes the internal representation of numbers from `Float` to `Double` throughout the Mica runtime. This provides greater precision for numerical operations and range definitions.

**Specific changes:**

-   **`TypeConverterHelper.kt`**: `String.toNumber()` now returns `Double`.
-   **`Variable.kt`**:
    -   `NumberVariable` now holds a `Double?`.
    -   `NumberRangeVariable` now holds a `ClosedDoubleRange?` (renamed from `ClosedFloatRange`).
    -   `IndefiniteNumberRangeVariable` now holds an `OpenEndDoubleRange?` (renamed from `OpenEndFloatRange`).
    -   The `create` factory method now reflects these changes.
-   **`ValueCoercionResolver.kt`**:
    -   Coercion from `Char` to `NumberType` now results in a `Double`.
    -   Coercion from `CharRange` to `NumberRangeType` now results in a `ClosedDoubleRange`.
    -   Coercion from `Number` to `CharType` and `BoolType` now operates on `Double`.
    -   Coercion from `NumberRange` to `CharRangeType` now operates on `ClosedDoubleRange`.
    -   `OpenEndFloatRange` is now `OpenEndDoubleRange`.
-   **`ValueComparatorResolver.kt`**: Comparisons for `NumberType`, `NumberRangeType`, and `IndefiniteNumberRangeType` now use `Double`, `ClosedDoubleRange`, and `OpenEndDoubleRange` respectively.
-   **`PrefixUnaryOperatorExpressionExecutor.kt`**: Unary `+` and `-` operators now operate on and return `Double`.
-   **`Comparators.kt`**:
    -   `compareTo` extension functions are updated for `ClosedDoubleRange` (renamed from `ClosedFloatRange`) and `OpenEndDoubleRange` (renamed from `OpenEndFloatRange`).
-   **`BinaryOperatorExpressionExecutor.kt`**:
    -   Arithmetic operations (`+`, `-`, `*`, `/`, `%`) now operate on and return `Double`.
    -   Range creation (`..`) for numbers now produces a `ClosedDoubleRange`.
-   **File Renames**:
    -   `ClosedFloatRange.kt` renamed to `ClosedDoubleRange.kt`.
    -   `OpenEndFloatRange.kt` renamed to `OpenEndDoubleRange.kt`.
-   **Class Renames**:
    -   `ClosedFloatRange` renamed to `ClosedDoubleRange`.
    -   `OpenEndFloatRange` renamed to `OpenEndDoubleRange`.
    These classes now use `Double` internally.

+++ mica/src/main/java/com/pointlessapps/granite/mica/runtime/executors/PrefixUnaryOperatorExpressionExecutor.kt
@@ -17,8 +17,8 @@ internal object PrefixUnaryOperatorExpressionExecutor {
 
         return when (expression.operatorToken.type) {
             Token.Operator.Type.Not -> !(value as Boolean)
-            Token.Operator.Type.Subtract -> -(value as Float)
-            Token.Operator.Type.Add -> value as Float
+            Token.Operator.Type.Subtract -> -(value as Double)
+            Token.Operator.Type.Add -> value as Double
             else -> throw RuntimeTypeException(
                 "Operator ${expression.operatorToken.type.valueLiteral()} is not applicable to ${
                     typeResolver.resolveExpressionType(expression).name
